[TAN-8637] Cover the events widget in one e2e suite instead of three
